feat: Reload Circle
There is now a reload circle that plays and goes around the crosshair based on how much time the reload takes for the gun. This is to help the player know their gun is reloading as without a reload animation it is difficult to see. Backgrounds for UI are now put in as well so it is not just floating pieces.