Feat: Crosshair maker, Sens Slider, Sound Slider, Custom settings
Custom settings such as crosshair settings, sound, sens are now all saved to a playercustomsettings class to allow for ease of starting back up. Player can now customize crosshairs in the options menu and get a preview in real time. Sens is now adjustable in menus and as well as a sound slider. Sound not implemented yet, but slider is there